+++ classes/AdminTabs/DashboardTab.php
@@ -6,6 +6,14 @@ use ProcessWire\Wire;
 
 class DashboardTab
 {
+    /**
+     * @param $form
+     * @param $data
+     * @return array
+     * @throws ProcessWire\Wire404Exception
+     * @throws ProcessWire\WireException
+     * @throws ProcessWire\WirePermissionException
+     */
     public function build($form, $data): array {
 
         $m         = \ProcessWire\wire('modules');